DEV-1449: schema updates for holdings/hathifile loading
* ocn uses bigint -- while the maximum real OCN fits comfortably in 32
  bits, it's possible we'll have other non-OCN numbers come through, and
  validation practices differ across different things (hathifiles,
  holdings, etc)

* index ocn and organization for holdings

+++ sql/007_oclc_concordance.sql
@@ -2,8 +2,8 @@ use `ht_repository`;
 
 DROP TABLE IF EXISTS `oclc_concordance`;
 CREATE TABLE `oclc_concordance` (
-  `oclc` int(10) unsigned NOT NULL,
-  `canonical` int(10) unsigned DEFAULT NULL,
+  `oclc` bigint unsigned NOT NULL,
+  `canonical` bigint unsigned DEFAULT NULL,
   PRIMARY KEY (`oclc`,`canonical`),
   INDEX (`canonical`)
 ) ENGINE=InnoDB
